[PHP] Show files-pull checkpoint lifecycle in tests
